// src/model/types/PayIDHeader.ts
export enum PayIDHeader {
    ALL = 'application/payid+json',
    XRPL_MAINNET = 'application/xrpl-mainnet+json',
    XRPL_DEVNET = 'application/xrpl-devnet+json',
    XRPL_TESTNET = 'application/xrpl-testnet+json',
    ACH = 'application/ach+json',
    BTC_MAINNET = 'application/btc-mainnet+json',
    BTC_TESTNET = 'application/btc-testnet+json',
    ETH_MAINNET = 'application/eth-mainnet+json'
}